frick

+++ src/main/java/baby/creed/modules/render/PurseSpoofer.java
@@ -4,11 +4,7 @@ import baby.creed.Creed;
 import baby.creed.events.PacketReceivedEvent;
 import baby.creed.modules.Module;
 import baby.creed.settings.NumberSetting;
-import baby.creed.util.PacketUtils;
 import net.minecraft.network.Packet;
-import net.minecraft.network.play.server.S3BPacketScoreboardObjective;
-import net.minecraft.network.play.server.S3CPacketUpdateScore;
-import net.minecraft.network.play.server.S3DPacketDisplayScoreboard;
 import net.minecraft.network.play.server.S3EPacketTeams;
 import net.minecraft.util.StringUtils;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
@@ -27,7 +23,6 @@ public class PurseSpoofer extends Module {
     public PurseSpoofer()
     {
         super("Purse Spoofer", Category.RENDER);
-
         addSettings(additionalCoins);
     }
 
@@ -43,30 +38,23 @@ public class PurseSpoofer extends Module {
         if (!isToggled() || Creed.mc.thePlayer == null || Creed.mc.theWorld == null) return;
 
         Packet<?> packet = event.packet;
-
         if (packet instanceof S3EPacketTeams)
         {
             S3EPacketTeams team = (S3EPacketTeams) packet;
-
             String strip = StringUtils.stripControlCodes(team.getPrefix()).toLowerCase();
 
             if (!strip.startsWith("purse: ")) return;
 
             final double purseValue = Double.parseDouble(strip.split(" ")[1].replaceAll(",", ""));
-
             final double addCoins = additionalCoins.getValue();
 
-            String newPurse = Creed.fancy + "fPurse: " + Creed.fancy + "6" + String.format("%,.1f", (purseValue + addCoins));
-
+            String newPurse = Creed.fancy + "fPurse: " + Creed.fancy + "6" + String.format("%,.1f", ((float)purseValue + (float)addCoins));
             System.out.println("found purse. New purse is " + newPurse);
-
             System.out.println("Values are purse: " + purseValue + " and add is " + addCoins);
 
             try {
                 Field field = S3EPacketTeams.class.getDeclaredField("prefix");
-
                 field.setAccessible(true);
-
                 field.set(team, newPurse);
             } catch (Exception e) {
                 e.printStackTrace();
